add workaround for randomly failing tests
On slower machines (such as the tests slaves of travis) the debounce
will sometimes not reduce all events of a unlink to one call. This is
can lead to weird effects as the first event is a 'change' which
triggers an unexpected read.

For now force a longer debounce when running tests