deprecate event emitter interface
It does not add much as reacting to changes is orthogonal to the
intended use case of the module. In addition the error event was a pain
as it forced you to set up a listener so that is already disabled.
`change` and `load` is still emitted but any use of `on` results in a
deprecation warning.

Fixes #6